build assigne select

+++ app/api/issues/[id]/route.ts
@@ -1,47 +1,51 @@
+import authOptions from "@/app/auth/authOptions";
 import { issueSchema } from "@/app/validationSchemas";
 import prisma from "@/prisma/client";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 
 interface Props {
-    params: { id: string}
+  params: { id: string };
 }
 
-export async function PATCH(request: NextRequest, { params} : Props) {
-    const body = await request.json();
-    const validation = issueSchema.safeParse(body);
-    if (!validation.success)
-        return NextResponse.json(validation.error.format(), { status: 400})
-    const issue = await prisma.issue.findUnique({
-        where: { id: parseInt(params.id)}
-    });
+export async function PATCH(request: NextRequest, { params }: Props) {
+  const session = await getServerSession(authOptions);
+  if (!session) return NextResponse.json({}, { status: 401 });
+  const body = await request.json();
+  const validation = issueSchema.safeParse(body);
+  if (!validation.success)
+    return NextResponse.json(validation.error.format(), { status: 400 });
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) },
+  });
 
-    if(!issue)
-        return NextResponse.json({error: 'Invalid issue'}, {status: 404})
+  if (!issue)
+    return NextResponse.json({ error: "Invalid issue" }, { status: 404 });
 
-    const updatedIssue = await prisma.issue.update({
-        where: { id: issue.id },
-        data: {
-            title: body.title,
-            description: body.description
-        }
-    });
+  const updatedIssue = await prisma.issue.update({
+    where: { id: issue.id },
+    data: {
+      title: body.title,
+      description: body.description,
+    },
+  });
 
-    return NextResponse.json(updatedIssue);
-
+  return NextResponse.json(updatedIssue);
 }
 
-export async function DELETE(request: NextRequest, { params} : Props) {
+export async function DELETE(request: NextRequest, { params }: Props) {
+  const session = await getServerSession(authOptions);
+  if (!session) return NextResponse.json({}, { status: 401 });
 
-    const issue = await prisma.issue.findUnique({
-        where: { id: parseInt(params.id) }
-    });
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) },
+  });
 
-    if(!issue)
-        return NextResponse.json({error: 'invalid issue', status: 404});
+  if (!issue) return NextResponse.json({ error: "invalid issue", status: 404 });
 
-        await prisma.issue.delete( {
-            where: { id: issue.id}
-        })
+  await prisma.issue.delete({
+    where: { id: issue.id },
+  });
 
-        return NextResponse.json({});
+  return NextResponse.json({});
 }
